4.18-4.21

+++ osa4/blogilista/controllers/blog.js
@@ -1,6 +1,7 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -17,11 +18,14 @@ blogRouter.get('/:id', async (request, response)=>{
     response.status(404).end()
   }
 })
-  
+
 blogRouter.post('/', async(request, response) => {
   const body = request.body
-
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -43,8 +47,17 @@ blogRouter.post('/', async(request, response) => {
 })
 
 blogRouter.delete('/:id', async(request, response)=>{
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const userid = decodedToken.id
+  const blog = await Blog.findById(request.params.id)
+
+  console.log(blog)
+
+  if ( blog.user.toString() === userid.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {return response.status(401).json({ error: 'Unauthorized' })}
 })
 
 blogRouter.put('/:id', async(request, response)=>{
